Added basic support for ANSI color schemes
Moved "searching" / API query logic to separate class
