+++ k8v/viewer.py
@@ -1,7 +1,11 @@
-import kubernetes
+import json
+
+from ansi.color import fg, bg
+import ansi.color.fx as fx
 
 import config
 from resource_types import ResourceType
+import searcher
 
 
 class Viewer:
@@ -11,57 +15,7 @@ class Viewer:
         self.config: config.Config = config
         self.delim: str = delim
 
-
-    def connect(self):
-        """Load the Kubernetes configuration and setup API endpoint connections."""
-        self.kubernetes_config = kubernetes.config.load_kube_config()
-        self.api_client = kubernetes.client.ApiClient(self.kubernetes_config)
-        self.api_core_v1 = kubernetes.client.CoreV1Api()
-        self.api_apps_v1 = kubernetes.client.AppsV1Api(self.api_client)
-        self.api_network_v1 = kubernetes.client.NetworkingV1Api()
-
-
-    def filter_resources(self, resources):
-        """Apply filtering logic to the specified resources."""
-
-        # fitler by "includes" first
-        for include in self.config.includes:
-            resources = filter(lambda x: include in x.metadata.name, resources)
-
-        # filter by label "selectors" next
-        for label,value in self.config.selectors.items():
-            resources = filter(lambda x: x.metadata.labels is not None and label in x.metadata.labels and x.metadata.labels[label] == value, resources)
-
-        # exclude anything undesirable lastly
-        for exclude in self.config.excludes:
-            resources = filter(lambda x: exclude not in x.metadata.name, resources)
-        return resources
-
-
-    def get_api_handler(self, type: ResourceType) -> str:
-        """Retrieve the API handler function to use for the specified namespace(s) and ResourceType."""
-        if type == ResourceType.CONFIG_MAP:
-            return self.api_core_v1.list_config_map_for_all_namespaces if self.config.namespaces is None else self.api_core_v1.list_namespaced_config_map
-        elif type == ResourceType.DEPLOYMENTS:
-            return self.api_apps_v1.list_deployment_for_all_namespaces if self.config.namespaces is None else self.api_apps_v1.list_namespaced_deployment
-        elif type == ResourceType.DAEMON_SETS:
-            return self.api_apps_v1.list_daemon_set_for_all_namespaces if self.config.namespaces is None else self.api_apps_v1.list_namespaced_daemon_set
-        elif type == ResourceType.REPLICA_SETS:
-            return self.api_apps_v1.list_replica_set_for_all_namespaces if self.config.namespaces is None else self.api_apps_v1.list_namespaced_replica_set
-        elif type == ResourceType.PODS:
-            return self.api_core_v1.list_pod_for_all_namespaces if self.config.namespaces is None else self.api_core_v1.list_namespaced_pod
-        elif type == ResourceType.SECRETS:
-            return self.api_core_v1.list_secret_for_all_namespaces if self.config.namespaces is None else self.api_core_v1.list_namespaced_secret
-        elif type == ResourceType.INGRESS:
-            return self.api_network_v1.list_ingress_for_all_namespaces if self.config.namespaces is None else self.api_network_v1.list_namespaced_ingress
-        elif type == ResourceType.PERSISTENT_VOLUME:
-            return self.api_core_v1.list_persistent_volume  if self.config.namespaces is None else None
-        elif type == ResourceType.PERSISTENT_VOLUME_CLAIM:
-            return self.api_core_v1.list_persistent_volume_claim_for_all_namespaces if self.config.namespaces is None else self.api_core_v1.list_namespaced_persistent_volume_claim
-        elif type == ResourceType.SERVICES:
-            return self.api_core_v1.list_service_for_all_namespaces if self.config.namespaces is None else self.api_core_v1.list_namespaced_service
-        else:
-            return None
+        self.searcher: searcher.Searcher = searcher.Searcher(self)
 
 
     def map_api_type(self, api_type: str) -> str:
@@ -90,44 +44,21 @@ class Viewer:
             return api_type
 
 
-    def get_pod_data(self, resource) -> [list,list]:
-        """Get any related configmap or secrets related to this resource."""
-
-        # search for "envFrom" sections in the container definitions
-        data: dict = { 'configmaps': [], 'secrets': [], 'pvcs': [], 'volumes': []}
-        if hasattr(resource, 'spec') and hasattr(resource.spec, 'containers'):
-            for container in resource.spec.containers:
-                if hasattr(container, 'env_from') and container.env_from is not None:
-                    for envFrom in container.env_from:
-                        if hasattr(envFrom, 'config_map_ref') and envFrom.config_map_ref is not None:
-                            data['configmaps'].append(envFrom.config_map_ref.name)
-                        if hasattr(envFrom, 'secret_ref') and envFrom.secret_ref is not None:
-                            data['secrets'].append(envFrom.secret_ref.name)
-
-        # search through "volume" definitions
-        if hasattr(resource, 'spec') and hasattr(resource.spec, 'volumes'):
-            for volume in resource.spec.volumes:
-                data['volumes'].append(volume)
-                if volume.config_map is not None:
-                    data['configmaps'].append(volume.config_map.name)
-                elif volume.secret is not None:
-                    data['secrets'].append(volume.secret.secret_name)
-                elif volume.persistent_volume_claim is not None:
-                    data['pvcs'].append(volume.persistent_volume_claim.claim_name)
-        return data
-
-
     def get_label_text(self, resource) -> str:
         """Get the text that should be dispalyed for labels in all resources."""
         if resource.metadata.labels is None:
             return ""
 
-        text: str = f"labels=["
-        labels: list = []
-        for label,value in resource.metadata.labels.items():
-            labels.append(f"{label}={value}")
-        text += f"{' '.join(labels)}] "
-        return text
+        message: str = self.get_ansi_text("attr2_name", "labels")
+        message += self.get_ansi_text("attr2_delim", "=[")
+        for num, (label,value) in enumerate(resource.metadata.labels.items()):
+            message += self.get_ansi_text("attr_name", label)
+            message += self.get_ansi_text("attr_delim", "=")
+            message += self.get_ansi_text("attr_value", value)
+            if num < len(resource.metadata.labels)-1:
+                message += ", "
+        message += self.get_ansi_text("attr2_delim", "] ")
+        return message
 
 
     def print(self, resource, type: ResourceType, delim: str = "") -> None:
@@ -140,9 +71,35 @@ class Viewer:
             self.print_full(resource, type, delim)
 
 
+    def get_ansi_text(self, key: str, text: str) -> str:
+        """Format a message with the specified text before resetting the ANSI code."""
+
+        # work with no schema selected
+        if self.colors == None:
+            return text
+
+        message = []
+        for data in self.colors[key]:
+            src = None
+            if data[0] == 'fg':
+                src = fg
+            elif data[0] == 'bg':
+                src = bg
+            elif data[0] == 'fx':
+                src = fx
+            message.append(getattr(src, data[1]))
+        message.append(text)
+        message.append(fx.reset)
+        return ''.join(map(str, message))
+
     def print_brief(self, resource, type: ResourceType, delim: str = "") -> None:
         """Print the **brief** display version of a resource."""
-        print(f"{delim}{self.map_api_type(resource.__class__.__name__)}/{resource.metadata.namespace}/{resource.metadata.name}")
+        message = delim + self.get_ansi_text("type", self.map_api_type(resource.__class__.__name__))
+        message += "/"
+        message += self.get_ansi_text("namespace", resource.metadata.namespace)
+        message += "/"
+        message += self.get_ansi_text("name", resource.metadata.name)
+        print(message)
 
     def print_default(self, resource, type: ResourceType, delim: str = "") -> None:
         """Print the **default** display version of a resource."""
@@ -158,10 +115,11 @@ class Viewer:
             extended_info += f"storage_class={resource.spec.storage_class_name} access_modes={resource.spec.access_modes} reclaim={resource.spec.persistent_volume_reclaim_policy} capacity={resource.spec.capacity['storage']}"
 
         elif type == ResourceType.PERSISTENT_VOLUME_CLAIM:
-            extended_info += f"storage_class={resource.spec.storage_class_name} access_modes={resource.spec.access_modes} capacity={resource.status.capacity['storage']} volume={resource.spec.volume_name} phase={resource.status.phase}"
+            extended_info += f"storage_class={resource.spec.storage_class_name} access_modes={resource.spec.access_modes} capacity={resource.status.capacity['storage']} "
+            extended_info += f"volume={resource.spec.volume_name} phase={resource.status.phase}"
 
         elif type == ResourceType.PODS:
-            pod_data = self.get_pod_data(resource)
+            pod_data = self.searcher.get_pod_data(resource)
             extended_info += f"config_maps={', '.join(pod_data['configmaps']) if len(pod_data['configmaps']) > 0 else ''} "
             extended_info += f"secrets={', '.join(pod_data['secrets']) if len(pod_data['secrets']) > 0 else ''} "
             extended_info += f"pvc={', '.join(pod_data['pvcs']) if len(pod_data['pvcs']) > 0 else ''}"
@@ -197,9 +155,17 @@ class Viewer:
         extended_info = '('+extended_info+')' if len(extended_info) > 0 else extended_info
         if post_info != "":
             post_info = "\n"+ post_info
-        print(f"{delim}{self.map_api_type(resource.__class__.__name__)}{'/'+resource.metadata.namespace if resource.metadata.namespace else ''}/{resource.metadata.name} {extended_info}{post_info}")
 
-        for related in self.search_for_related(resource, type):
+        message = delim + self.get_ansi_text("type", self.map_api_type(resource.__class__.__name__))
+        message += "/"
+        if resource.metadata.namespace:
+            message += self.get_ansi_text("namespace", resource.metadata.namespace) 
+            message += "/"
+        message += self.get_ansi_text("name", resource.metadata.name)
+        message += " "
+        print(f"{message}{extended_info}{post_info}")
+
+        for related in self.searcher.search_for_related(resource, type):
             if type == ResourceType.DEPLOYMENTS:
                 self.print(related, ResourceType.REPLICA_SETS, delim=delim+self.delim)
             elif type == ResourceType.DAEMON_SETS:
@@ -212,76 +178,24 @@ class Viewer:
         self.print_default(resource, type, delim)
 
 
-    def search_for_related(self, resource, type: ResourceType) -> list:
-        """Search for any related resources of the given type."""
-        resources: list = []
-        label_expr: str = ""
-
-        if hasattr(resource, 'spec') and hasattr(resource.spec, 'selector'):
-            selector = resource.spec.selector
-            if hasattr(selector, 'match_expressions') and selector.match_expressions is not None and len(selector.match_expressions) > 0:
-                pass # ignore since we cannot easily evaluate right now
-            if hasattr(selector, 'match_labels'):
-                label_count: int = len(selector.match_labels)-1
-                for num, label in enumerate(selector.match_labels.items()):
-                    label_expr += label[0] +"="+ label[1]
-                    if num < label_count:
-                        label_expr += ","
-        elif resource.metadata.labels is not None and len(resource.metadata.labels) > 0:
-            for num, label in enumerate(resource.metadata.labels):
-                label_expr += label +"="+ resource.metadata.labels[label]
-                if num < len(resource.metadata.labels)-1:
-                    label_expr += ","
-
-        if type == ResourceType.DEPLOYMENTS:
-            resources = self.search(ResourceType.REPLICA_SETS, label_selector=label_expr)
-        elif type == ResourceType.DAEMON_SETS:
-            resources = self.search(ResourceType.PODS, label_selector=label_expr)
-        elif type == ResourceType.REPLICA_SETS:
-            resources = self.search(ResourceType.PODS, label_selector=label_expr)
-
-        return resources
-
-
-    def search(self, type: ResourceType, **kwargs) -> list:
-        """Search for matching resources for the specified type."""
-        resources = []
-
-        # deterine which API call to use
-        handler = self.get_api_handler(type)
-        if handler is None:
-            return resources 
-            
-        try:
-            if self.config.verbose:
-                print(f"Searching for {type.value[0]}") 
-
-            if self.config.namespaces is None:
-                api_response = handler(**kwargs)
-                for d in api_response.items:
-                    resources.append(d)
-            else:
-                for ns in self.config.namespaces:
-                    api_response = handler(ns, **kwargs)
-                    for d in api_response.items:
-                        resources.append(d)
-        except Exception as e:
-            print(f"Exeception occurred while searching for resources ({type.value[0]}): {e}")
-            raise e
-
-        # apply filtering logic to matching resources
-        return sorted(resources, key=lambda x: x.metadata.name)
-
-
     def view(self) -> None:
         """Use the input parameters to create a View of the desired resources and their relationships."""
+        try:
+            schemes = json.load(open("k8v/color-schemes.json"))["schemes"]
+            if self.config.colors in schemes:
+                self.colors = schemes[self.config.colors]
+            else:
+                self.colors = None
+        except Exception as e:
+            print(f"Exception occurred loading color schemes: {e}")
+            raise e
 
         # show the input parameters
         if self.config.verbose:
             print(f"Display mode={self.config.display_type}, namespaces={self.config.namespaces}, resources={self.config.resources}, filters={self.config.includes}, selectors={self.config.selectors}")
 
         # setup the API handlers
-        self.connect()
+        self.searcher.connect()
 
         # setup default namespace if no overrides specified
         if self.config.namespaces is not None and len(self.config.namespaces) == 0:
@@ -304,6 +218,6 @@ class Viewer:
         # search for matching (and filtered) resources and print them out
         # using the desired display_mode.
         for type in self.config.resources:
-            resources = self.search(type)
-            for resource in self.filter_resources(resources):
+            resources = self.searcher.search(type)
+            for resource in self.searcher.filter_resources(resources):
                 self.print(resource, type)
